fix: improve build.sh script

+++ build.sh
@@ -6,7 +6,11 @@ env GOOS=darwin GOARCH=arm64 go build -o ./dist/protoc-gen-tsx.darwin.arm64
 env GOOS=linux GOARCH=amd64 go build -o ./dist/protoc-gen-tsx.linux.amd64
 env GOOS=linux GOARCH=arm64 go build -o ./dist/protoc-gen-tsx.linux.arm64
 
-ln -s ./dist/protoc-gen-tsx.linux.arm64 ./dist/protoc-gen-tsx
+if [ ! -f ./dist/protoc-gen-tsx ] && [ ! -L ./dist/protoc-gen-tsx ]; then
+  ln -s ./dist/protoc-gen-tsx.linux.arm64 ./dist/protoc-gen-tsx
+fi
+
+chmod a+x ./dist/*
 
 if [ "${CI}" != "true" ]; then
     # Stripping symbols to make the binary more lightweight
